fixed story, moved position handler to helper
