+++ src/components/extendedTooltip/ExtendedTooltip.stories.mdx
@@ -16,8 +16,6 @@ import { SVGIcon } from "../svgicon";
 > - `showTimeout`: `number` - **optional** (delay time in miliseconds for opening of tooltip)
 > - `hideTimeout`: `number` - **optional** (delay time in miliseconds for closing of tooltip)
 
-<br style={{ marginBottom: "2rem" }} />
-
 ## top
 
 <Canvas>
@@ -216,24 +214,22 @@ import { SVGIcon } from "../svgicon";
     >
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
+        placement="left"
         position="start"
-        placement="left"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
-
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
+        placement="left"
         position="center"
-        placement="left"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
-
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
-        position="end"
         placement="left"
+        position="end"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
@@ -316,24 +312,22 @@ import { SVGIcon } from "../svgicon";
     >
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
+        placement="right"
         position="start"
-        placement="right"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
-
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
+        placement="right"
         position="center"
-        placement="right"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
-
       <ExtendedTooltip
         caption="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vestibulum sollicitudin."
-        position="end"
         placement="right"
+        position="end"
       >
         <button type="button">Hover Me!</button>
       </ExtendedTooltip>
